+++ packages/lucia-sveltekit/src/auth/index.ts
@@ -24,7 +24,13 @@ import {
 } from "./session.js";
 import { handleServerSessionFunction } from "./load.js";
 import clc from "cli-color";
-import { Adapter, UserData, UserSchema } from "../adapter/index.js";
+import {
+    Adapter,
+    SessionAdapter,
+    UserAdapter,
+    UserData,
+    UserSchema,
+} from "../adapter/index.js";
 
 export const lucia = <C extends Configurations>(configs: C) => {
     return new Auth(configs) as Omit<Auth<C>, "getAuthSession">;
@@ -52,7 +58,10 @@ export class Auth<C extends Configurations = any> {
         };
         this.context = {
             auth: this,
-            adapter: configs.adapter,
+            adapter:
+                "user" in configs.adapter
+                    ? { ...configs.adapter.user, ...configs.adapter.session }
+                    : configs.adapter,
             generateCustomUserId:
                 configs.generateCustomUserId || (async () => null),
             env: configs.env,
@@ -142,7 +151,12 @@ export class Auth<C extends Configurations = any> {
 }
 
 interface Configurations {
-    adapter: Adapter;
+    adapter:
+        | Adapter
+        | {
+              user: UserAdapter;
+              session: SessionAdapter;
+          };
     env: Env;
     generateCustomUserId?: () => Promise<string | null>;
     csrfProtection?: boolean;